refactor: refactor code, remove users content, move the users logic to the search result, remove the bug of increasing the screen size in the safary browser on a sick device
