+++ index.html
@@ -4,6 +4,13 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Задание для стажировки</title>
+    <style>
+         @media screen and (-webkit-min-device-pixel-ratio: 0) {
+                       body input:focus {
+                               font-size: 16px;
+                           }
+                       }
+    </style>
   </head>
   <body>
     <div id="root"></div>
