+++ src/App.tsx
@@ -2,13 +2,15 @@ import React from "react";
 
 import {SearchForm} from "./components/SearchFrom/SearchForm";
 import {SearchResults} from "./components/SearchResults/SearchResults";
-import {ContextProvider} from "./utilities";
+import {SearchFormProvider} from "./utilities";
+
+/*import "./styles.css";*/
 
 export const App: React.FC = () => {
   return (
-    <ContextProvider>
+    <SearchFormProvider>
       <SearchForm/>
       <SearchResults/>
-    </ContextProvider>
+    </SearchFormProvider>
   );
 }
