+++ src/components/SearchFrom/styles.css
@@ -23,3 +23,9 @@
 .searchForm input[type="submit"] {
     width: 150px;
 }
+
+@media screen and (-webkit-min-device-pixel-ratio: 0) {
+    select:focus, textarea:focus, input:focus {
+        font-size: 16px;
+    }
+}
