+++ src/components/SearchResults/SearchResults.tsx
@@ -2,20 +2,24 @@ import React from "react";
 
 import {Loader} from "../Loader/Loader.tsx";
 import {UserCard} from "../UserCard/UserCard";
-import {useUsersContext} from "../../utilities";
 
 import "./styles.css";
+import {IUsers, useDebounce, useFetch, useSearchFormContext} from "../../utilities";
+
+const API_URL = "https://dummyjson.com/users";
 
 export const SearchResults: React.FC = () => {
-  const {users, error, isLoading} = useUsersContext()
+  const {value} = useSearchFormContext()
+  const debouncedSearchTerm = useDebounce({value, delay: 500});
+  const {data, error, isLoading} = useFetch<IUsers>(`${API_URL}/search?q=${debouncedSearchTerm}`)
 
   if (isLoading) return <Loader/>
   if (error) return <span className="usersError">{error}</span>
-  if (!users.length) return <span className="usersError">Пользатель не найден!</span>
+  if (!data?.users.length) return <span className="usersError">Пользатель не найден!</span>
 
   return (
     <div className="usersList">
-      {users.map((user) => (
+      {data?.users.map((user) => (
         <UserCard key={user.id} {...user}  />
       ))}
     </div>
