+++ src/components/SearchResults/styles.css
@@ -10,3 +10,9 @@
     justify-content: center;
     color: red;
 }
+
+@media screen and (-webkit-min-device-pixel-ratio: 0) {
+    select:focus, textarea:focus, input:focus {
+        font-size: 16px;
+    }
+}
