+++ src/utilities/context/index.ts
@@ -1,2 +1,2 @@
-export {ContextProvider} from './ContextProvider.tsx'
-export {useUsersContext, useSearchFormContext} from './useContext.ts'
+export {SearchFormProvider} from './SearchFormProvider.tsx'
+export {useSearchFormContext} from './useContext.ts'
