+++ src/utilities/context/useContext.ts
@@ -1,10 +1,5 @@
 import {useContext, useMemo} from "react";
-import {SearchFormContext, UsersContext} from "./ContextProvider.tsx";
-
-export const useUsersContext = () => {
-  const {users, error, isLoading} = useContext(UsersContext)
-  return useMemo(() => ({users, error, isLoading}), [error, isLoading, users])
-}
+import {SearchFormContext} from "./SearchFormProvider.tsx";
 
 export const useSearchFormContext = () => {
   const {value, setValue} = useContext(SearchFormContext)
